+++ app/Http/Controllers/PostController.php
@@ -9,7 +9,12 @@ use Illuminate\Support\Facades\Auth;
 class PostController extends Controller
 {
    
-
+    public function index()
+    {
+        $posts = Post::paginate(5);
+        return view('posts', ['posts' => $posts]);
+    }
+    
 
     /**
      * Store a newly created resource in storage.
@@ -59,6 +64,11 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        // on met en place la verification avec la ,policy
+        // pour etre sur que l'utilisateur a le droit d'effectuer l'action
+
+        $this->authorize('update', $post);
+
         // 1) on valide les champs en précisant les critères attendus
         // (memes critères que dans la fonction store)
         $request->validate([
@@ -82,6 +92,11 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+                // on met en place la verification avec la ,policy
+        // pour etre sur que l'utilisateur a le droit d'effectuer l'action
+
+        $this->authorize('delete', $post);
+
         // je supprime le message
         $post->delete();
 
@@ -100,9 +115,8 @@ class PostController extends Controller
 
         // on va chercher les messages qui comportent cette recherche
         // dans leur tags et / ou dans leur contenu
-        $posts = Post::where('posts', 'tags', 'like', "%$search%") //1er critère: la chaine recherchée est dans les tags
-        ->orwhere('posts.content', 'like', "%$search%") // 2eme critère: la chaine recherchée est dans le contenu
-        ->with('comments') // EAGER LOADING pour charger les relations necessaires
+        $posts = Post::where('tags', 'like', "%$search%") //1er critère: la chaine recherchée est dans les tags
+        ->orwhere('content', 'like', "%$search%") // 2eme critère: la chaine recherchée est dans le contenu
         ->latest()->paginate(3); // comments.user = nested eager loading (eager loading en cascade)
 
         return view('home', ['posts' => $posts]);
